+++ services/mailservice/inbucket.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-package utils
+package mailservice
 
 import (
 	"bytes"
